✨ model controller machine 생성
모델 컨트롤러 머신 생성
